Fixed UrlBuilder prepending an extra & if parameters are empty (#72)
I noted that the UrlBuilder always prepends & to the checksum parameter even if the parameters. The final URL then looks like [...]/getMeetings?&chkecksum=[...]. This breaks nothing but is a bit confusing.

Therefor: Fixed and test case added. As I am currently working with PHP 8, this is based on the PHP 8 PR (#70) and a for now a draft for this reason.

Co-authored-by: Samuel Weirich <samuel.weirich@gmx.de>